+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 This file contains the list of changes made to the JLS project.
 
 
+## 0.2.1
+
+2021 Mar 8
+
+*   Fixed timeout on close when bursting data.
+*   Added annotate_stress example.
+*   Added flush.
+
+
 ## 0.2.0
 
 2021 Mar 8
@@ -12,7 +21,7 @@ This file contains the list of changes made to the JLS project.
 *   Fixed example/generate.py length argument.
 *   Updated numpy dependency to 1.20.
 *   Fixed pyjls console_script.
-*   Changed logging interface and connected natvie to python logging.
+*   Changed logging interface and connected native to python logging.
 
 
 ## 0.1.0